requirements.txt
# Core dependencies
flask==3.0.0
python-dotenv==1.0.0

# AI & ML
openai==1.12.0
sentence-transformers==2.3.1
scikit-learn==1.4.0
numpy==1.26.3

# Utilities
werkzeug==3.0.1
